requirements.txt
asgiref==3.5.2
Django==4.0.5
django-debug-toolbar==3.4.0
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
